split disk-selector into persistent-disk-selector and standard-disk-selector
